refactor: move RightSection
